+++ wger/exercises/views/equipment.py
@@ -61,7 +61,7 @@ class EquipmentListView(LoginRequiredMixin, PermissionRequiredMixin, ListView):
     """
 
     model = Equipment
-    fields = '__all__'
+    fields = ['name']
     template_name = 'equipment/admin-overview.html'
     context_object_name = 'equipment_list'
     paginate_by = PAGINATION_OBJECTS_PER_PAGE