+++ wger/urls.py
@@ -21,7 +21,10 @@ from django.conf.urls import include
 from django.conf.urls.i18n import i18n_patterns
 from django.conf.urls.static import static
 from django.contrib import admin
-from django.contrib.sitemaps.views import sitemap
+from django.contrib.sitemaps.views import (
+    index,
+    sitemap
+)
 from django.urls import path
 
 # Third Party
@@ -51,25 +54,24 @@ router = routers.DefaultRouter()
 #
 
 # Manager app
-router.register(r'workout', manager_api_views.WorkoutViewSet, basename='workout')
-router.register(r'workoutsession', manager_api_views.WorkoutSessionViewSet, basename='workoutsession')
-router.register(r'schedulestep', manager_api_views.ScheduleStepViewSet, basename='schedulestep')
-router.register(r'schedule', manager_api_views.ScheduleViewSet, basename='schedule')
 router.register(r'day', manager_api_views.DayViewSet, basename='day')
 router.register(r'set', manager_api_views.SetViewSet, basename='Set')
 router.register(r'setting', manager_api_views.SettingViewSet, basename='Setting')
+router.register(r'workout', manager_api_views.WorkoutViewSet, basename='workout')
+router.register(r'workoutsession', manager_api_views.WorkoutSessionViewSet, basename='workoutsession')
 router.register(r'workoutlog', manager_api_views.WorkoutLogViewSet, basename='workoutlog')
+router.register(r'schedulestep', manager_api_views.ScheduleStepViewSet, basename='schedulestep')
+router.register(r'schedule', manager_api_views.ScheduleViewSet, basename='schedule')
 
 # Core app
-router.register(r'userprofile', core_api_views.UserProfileViewSet, basename='userprofile')
-router.register(r'language', core_api_views.LanguageViewSet, basename='language')
 router.register(r'daysofweek', core_api_views.DaysOfWeekViewSet, basename='daysofweek')
+router.register(r'language', core_api_views.LanguageViewSet, basename='language')
 router.register(r'license', core_api_views.LicenseViewSet, basename='license')
+router.register(r'userprofile', core_api_views.UserProfileViewSet, basename='userprofile')
 router.register(r'setting-repetitionunit', core_api_views.RepetitionUnitViewSet, basename='setting-repetition-unit')
 router.register(r'setting-weightunit', core_api_views.WeightUnitViewSet, basename='setting-weight-unit')
 
 # Exercises app
-# Add router for viewing exercise info
 router.register(r'exerciseinfo', exercises_api_views.ExerciseInfoViewset, basename='exerciseinfo')
 router.register(r'exercise', exercises_api_views.ExerciseViewSet, basename='exercise')
 router.register(r'equipment', exercises_api_views.EquipmentViewSet, basename='api')
@@ -85,6 +87,7 @@ router.register(r'weightunit', nutrition_api_views.WeightUnitViewSet, basename='
 router.register(r'ingredientweightunit', nutrition_api_views.IngredientWeightUnitViewSet, basename='ingredientweightunit')
 router.register(r'nutritionplan', nutrition_api_views.NutritionPlanViewSet, basename='nutritionplan')
 router.register(r'nutritionplaninfo', nutrition_api_views.NutritionPlanInfoViewSet, basename='nutritionplaninfo')
+router.register(r'nutritiondiary', nutrition_api_views.LogItemViewSet, basename='nutritiondiary')
 router.register(r'meal', nutrition_api_views.MealViewSet, basename='meal')
 router.register(r'mealitem', nutrition_api_views.MealItemViewSet, basename='mealitem')
 
@@ -111,10 +114,10 @@ urlpatterns = i18n_patterns(
     path('config/', include(('wger.config.urls', 'config'), namespace='config')),
     path('gym/', include(('wger.gym.urls', 'gym'), namespace='gym')),
     path('email', include(('wger.mailer.urls', 'email'), namespace='email')),
-    path('sitemap.xml',
-         sitemap,
-         {'sitemaps': sitemaps},
-         name='sitemap')
+    path('sitemap.xml', index, {'sitemaps': sitemaps}, name='sitemap'),
+    path('sitemap-<section>.xml',
+         sitemap, {'sitemaps': sitemaps},
+         name='django.contrib.sitemaps.views.sitemap')
 )
 
 #